requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
fastapi==0.135.3
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
isodate==0.7.2
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
platformdirs==4.9.4
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
pwdlib==0.3.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
requests-file==3.0.1
requests-toolbelt==1.0.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.43.0
watchfiles==1.1.1
websockets==16.0
zeep==4.3.2
